Moved basic tests back into organization test, finished refactoring testCascade. All tests passing
